revamped ecohydrology_gridhydromet
- updated the ecohydrology_gridhydromet folder with the latest files and
file structure.
